+++ Age of war/src/UserInterface/Clickable.java
@@ -2,5 +2,5 @@ package UserInterface;
 
 public interface Clickable {
 	public void Click();
-	public boolean isClicked(double X, double Y);
+	public boolean isClicked(float X, float Y);
 }
